LayoutTests/storage/indexeddb/properties-disabled-at-runtime-expected.txt

ALERT: undefined
ALERT: 1
ALERT: undefined
ALERT: 2
ALERT: undefined
ALERT: 3
ALERT: undefined
ALERT: 4
ALERT: undefined
ALERT: 5
ALERT: undefined
ALERT: 6
ALERT: undefined
ALERT: 7
ALERT: undefined
ALERT: 8
ALERT: undefined
ALERT: 9
ALERT: undefined
ALERT: 10
ALERT: undefined
ALERT: 11
ALERT: undefined
ALERT: 12
ALERT: undefined
ALERT: 13
ALERT: undefined
ALERT: 14
ALERT: undefined
ALERT: 15
ALERT: undefined
ALERT: 16
ALERT: undefined
ALERT: 17
ALERT: undefined
ALERT: 18
This test makes sure objects/properties related to IndexedDatabase are both undefined and read/write when the feature is disabled at runtime.
